[Event] Rename Dispatcher* to EventDispatcher*.
