+++ .github/ci/psalm/psalm-baseline.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <files psalm-version="6.15.1@28dc127af1b5aecd52314f6f645bafc10d0e11f9">
-  <file src="../../../src/Valkyrja/Event/Dispatcher/Dispatcher.php">
+  <file src="../../../src/Valkyrja/Event/Dispatcher/EventDispatcher.php">
     <MixedMethodCall>
       <code><![CDATA[new $eventId()]]></code>
     </MixedMethodCall>
